docs(calendar): split stickersheet for single/range
