+++ packages/components/src/Calendar/CalendarSingle/_docs/CalendarSingle.stickersheet.stories.tsx
@@ -5,11 +5,11 @@ import {
   StickerSheet,
   StickerSheetStory,
 } from "~storybook/components/StickerSheet"
-import styles from "../Calendar.module.scss"
-import { CalendarRange, CalendarSingle, CalendarSingleProps } from "../index"
+import styles from "../../Calendar.module.scss"
+import { CalendarSingle, CalendarSingleProps } from "../index"
 
 export default {
-  title: "Components/Date controls/Calendars",
+  title: "Components/Date controls/Calendars/CalendarSingle",
   parameters: {
     chromatic: { disable: false },
     controls: { disable: true },
@@ -89,31 +89,6 @@ const StickerSheetTemplate: StickerSheetStory = {
           <StickerSheet.Row>
             <CalendarSingleExample id="id--calendar-navigation--hover" />
             <CalendarSingleExample id="id--calendar-navigation--focus" />
-          </StickerSheet.Row>
-        </StickerSheet.Body>
-      </StickerSheet>
-
-      <StickerSheet heading="Calendar Range">
-        <StickerSheet.Body>
-          <StickerSheet.Row rowTitle="Default">
-            <CalendarRange
-              selected={{
-                from: new Date("2022-02-19"),
-                to: new Date("2022-03-04"),
-              }}
-            />
-          </StickerSheet.Row>
-
-          <StickerSheet.Row rowTitle="With divider">
-            <div style={{ padding: "1.5rem 0" }}>
-              <CalendarRange
-                selected={{
-                  from: new Date("2022-02-19"),
-                  to: new Date("2022-03-04"),
-                }}
-                hasDivider
-              />
-            </div>
           </StickerSheet.Row>
         </StickerSheet.Body>
       </StickerSheet>
